backend/requirements.txt
fastapi>=0.123.0
uvicorn[standard]>=0.37.0
websockets>=13.0.1
pydantic>=2.12.0
pydantic-settings>=2.12.0
httpx>=0.28.0
python-dotenv>=1.0.0
